Break voting logic out into package.
