+++ contracts/cw-govmod-single/Cargo.toml
@@ -41,7 +41,8 @@ serde = { version = "1.0", default-features = false, features = ["derive"] }
 thiserror = { version = "1.0" }
 cw-governance = { path = "../cw-governance", version = "0.1.0", features = ["library"] }
 cw-governance-macros = { path = "../../packages/cw-governance-macros", version = "0.1.0" }
-cw-governance-interface = { version = "0.1.0", path = "../../packages/cw-governance-interface" } 
+cw-governance-interface = { version = "0.1.0", path = "../../packages/cw-governance-interface" }
+voting = { version = "0.1.0", path = "../../packages/voting" }
 
 [dev-dependencies]
 cosmwasm-schema = { version = "1.0.0-beta5" }
