+++ contracts/cw-govmod-single/src/msg.rs
@@ -3,7 +3,7 @@ use cw_utils::{Duration, Expiration};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
-use crate::{proposal::Vote, threshold::Threshold};
+use crate::threshold::Threshold;
 use cw_governance_macros::govmod_query;
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
@@ -74,7 +74,7 @@ pub enum ExecuteMsg {
         /// The ID of the proposal to vote on.
         proposal_id: u64,
         /// The senders position on the proposal.
-        vote: Vote,
+        vote: voting::Vote,
     },
     /// Causes the messages associated with a passed proposal to be
     /// executed by the DAO.
