+++ contracts/cw-govmod-sudo/Cargo.toml
@@ -38,7 +38,7 @@ serde = { version = "1.0", default-features = false, features = ["derive"] }
 thiserror = { version = "1.0" }
 cw-governance = { path = "../cw-governance", version = "0.1.0", features = ["library"] }
 cw-governance-macros = { version = "0.1.0", path = "../../packages/cw-governance-macros" }
-cw-governance-interface = { version = "0.1.0", path = "../../packages/cw-governance-interface" } 
+cw-governance-interface = { version = "0.1.0", path = "../../packages/cw-governance-interface" }
 
 [dev-dependencies]
 cosmwasm-schema = { version = "1.0.0-beta5" }
