+++ packages/voting/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "voting"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+cosmwasm-std = { version = "1.0.0-beta5" }
+schemars = "0.8"
+serde = { version = "1.0", default-features = false, features = ["derive"] }
